% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsplot.R
\name{jsplot.default}
\alias{jsplot.default}
\title{jsplot.default - error handling for creation of jsplot for non-dataframe type of object}
\usage{
\method{jsplot}{default}(data, ..., environment = parent.frame())
}
\description{
jsplot.default - error handling for creation of jsplot for non-dataframe type of object
}

