% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsplot.R
\name{jsplot}
\alias{jsplot}
\title{jsplot - a class that contains information necessary to render
    a base plot in javascript.}
\usage{
jsplot(data, ..., environment = parent.frame())
}
\arguments{
\item{data}{data frame that contains the dataset wish to be plotted. Defaults to NULL.}
}
\value{
A jsplot object that represents a base plot to the dataset 
        originally passed in as an argument.
}
\description{
jsplot - a class that contains information necessary to render
    a base plot in javascript.
}

