% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{render}
\alias{render}
\title{render - renders the jsplot object into html widget output}
\usage{
render(jsplot, width = NULL, height = NULL, elementId = NULL)
}
\description{
render accepts a layered jsplot and render a svg plot with specified
layer using other jsfunctions.
}

