% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sunburst.R
\name{sunburst}
\alias{sunburst}
\title{Sunburst - a d3 implementation of sunburst chart in R.}
\usage{
sunburst(src, width = NULL, height = NULL, elementId = NULL)
}
\description{
Sunburst - a d3 implementation of sunburst chart in R.
}

